requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
connexion[swagger-ui]==3.0.5
connexion[flask]==3.0.5
connexion[uvicorn]==3.0.5
pymongo==4.6.1
pyyaml-include==1.3.2

